Reorganize some tuple-related functionality beneath agency/tuple/detail
