+++ agency/detail/control_structures/executor_functions/shared_parameter_packaging.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <agency/detail/tuple_utility.hpp>
+#include <agency/tuple/detail/tuple_utility.hpp>
 #include <agency/detail/integer_sequence.hpp>
 #include <agency/tuple.hpp>
 #include <agency/detail/tuple_matrix.hpp>
